k256: factor away `Scalar::from_overflow`
This was a private method used to construct a `Scalar` from a
`WideScalar`.

Instead, the previous logic is reimplemented in terms of `UInt`, with
the implementation factored into the `WideScalar::reduce` methods(s).
