--- k256/src/arithmetic/util.rs
@@ -1,44 +1,0 @@
-//! Helper functions.
-// TODO(tarcieri): replace these with `crypto-bigint`
-
-#[cfg(test)]
-use num_bigint::{BigUint, ToBigUint};
-#[cfg(test)]
-use num_traits::cast::ToPrimitive;
-
-/// Computes `a - (b + borrow)`, returning the result along with the new borrow. 32-bit version.
-#[cfg(target_pointer_width = "32")]
-#[inline(always)]
-pub const fn sbb(a: u32, b: u32, borrow: u32) -> (u32, u32) {
-    let ret = (a as u64).wrapping_sub((b as u64) + ((borrow >> 31) as u64));
-    (ret as u32, (ret >> 32) as u32)
-}
-
-/// Computes `a - (b + borrow)`, returning the result along with the new borrow. 64-bit version.
-#[cfg(target_pointer_width = "64")]
-#[inline(always)]
-pub const fn sbb(a: u64, b: u64, borrow: u64) -> (u64, u64) {
-    let ret = (a as u128).wrapping_sub((b as u128) + ((borrow >> 63) as u128));
-    (ret as u64, (ret >> 64) as u64)
-}
-
-/// Converts a byte array (big-endian) to BigUint.
-#[cfg(test)]
-pub fn bytes_to_biguint(bytes: &[u8; 32]) -> BigUint {
-    bytes
-        .iter()
-        .enumerate()
-        .map(|(i, w)| w.to_biguint().unwrap() << ((31 - i) * 8))
-        .sum()
-}
-
-/// Converts a BigUint to a byte array (big-endian).
-#[cfg(test)]
-pub fn biguint_to_bytes(x: &BigUint) -> [u8; 32] {
-    let mask = BigUint::from(u8::MAX);
-    let mut bytes = [0u8; 32];
-    for i in 0..32 {
-        bytes[i] = ((x >> ((31 - i) * 8)) as BigUint & &mask).to_u8().unwrap();
-    }
-    bytes
-}
